[wip] improving file structure. removing minified files from git. added a 'npm run dev' script. :sushi: I am hungry
