+++ .gitignore
@@ -1,2 +1,2 @@
 node_modules
-client/dist/bundle.js
+/dist/*.min.js
