+++ package.json
@@ -5,6 +5,7 @@
   "private": true,
   "scripts": {
     "build": "NODE_ENV=production webpack --config webpack.config.js --progress --colors",
+    "dev": "webpack --config webpack.config.js --progress --colors --watch",
     "test": "tape --require babel-register tests/*.tests.js | faucet",
     "start": "node ./app.js"
   },
