--- src/components/component.js
@@ -1,6 +1,0 @@
-// Sample A-Frame component, a.k.a, HTML tag attribute, creation.
-AFRAME.registerComponent('hello-world', {
-    init: function () {
-        console.log('Hello, World!');
-    }
-});
